feat: get the content from markdown files in informasi-terbaru directory
